calc_corr {hmsPM}	R Documentation

Calculate Correlation/Contingency Matrix

Description

Calculate Correlation/Contingency Matrix

Usage

calc_corr(df, feature_scale = NULL, cutoff = 0, method = "spearman",
  verbose = TRUE)

Arguments

df
Data frame with variables of one scale level / type.

feature_scale
Scale level of variables.

cutoff
Lower cutoff level. Value in [0,1]. Remove rows and columns in which no correlation value is above the cutoff value.

method
Correlation method. Options: 'spearman' or 'pearson'

verbose
(Logical) Show messages?

Value

Correlation/Contingency matrix.

Examples

TODO

[Package hmsPM version 1.0.0 Index]