calc_individual_dependencies {hmsPM}	R Documentation

Calculate individual dependance

Description

Calculate individual dependance

Usage

calc_individual_dependencies(df_data, df_data_orig = NULL, fit,
  formula_string, sparse = TRUE, feature_names = NULL,
  b_sample = NULL, b_all = NULL, id_name = "id", quantiles = seq(0,
  1, 0.1), verbose = TRUE)

Arguments

df_data
Data frame for which individual dependencies should be calculated.

df_data_orig
Original data frame with data (needed to derive quantiles). If NULL, df_data is taken.

fit
Caret fit.

formula_string
Formula string used in caret fit (or for creation of matrix used).

sparse
Use sparse matrix for caret fit if TRUE. If NULL then no matrix is created and df_data is passed.

feature_names
Names of features to be calculated. If NULL all features are used.

b_sample
Named vector with priori probabilities of sampled data.

b_all
Named vector with priori probabilities of all data.

id_name
Name of id Variable.

quantiles
Numeric vector with probs of quantiles used for metric feature plot points.

verbose
(Logical) Show messages?

Value

Data frame with partial dependency infos.

Examples

TODO

[Package hmsPM version 1.0.0 Index]