calc_learningcurve {hmsPM}	R Documentation

Calculate Learning Curve

Description

Calculate Learning Curve

Usage

calc_learningcurve(df_data, formula_string, sparse = TRUE, method,
  tune_grid, chunks_pct = c(seq(1, 10, 1), seq(20, 100, 10)),
  balanced = TRUE, metric = "AUC", verbose = TRUE)

Arguments

df_data
Data frame for caret fit. Must contain variable fold with members train and test.

formula_string
Formula to use in caret fit (or for creation of matrix used).

sparse
Use sparse matrix for caret fit if TRUE. If NULL then no matrix is created and df_data is passed.

method
Method to use in caret fit.

tune_grid
Tune grid to use in caret fit.

chunks_pct
Percentages of training data to use for fits.

balanced
Balance the target as long as possible if TRUE.

metric
Tuning metric.

verbose
(Logical) Show messages?

Value

Data frame with learning curve metrics.

See Also

Other learning curve functions: plot_learningcurve

Examples

TODO

[Package hmsPM version 1.0.0 Index]