calc_simulation {hmsPM}	R Documentation

Run Simulation

Description

Run Simulation

Usage

calc_simulation(df_data, n_sim = 5, metric, sample_frac_train = 0.5,
  sample_frac_test = 0.5, l_methods)

Arguments

df_data
Data frame for caret fit. Must contain variable "fold" with members "train" and "test".

n_sim
Number of simulations to run.

metric
Tuning metric.

sample_frac_train
Fraction of training data to be sampled in each run.

sample_frac_test
Fraction of test data to be sampled in each run.

l_methods
List of methods. Each method itself is represented by a named list comprising the elements method, formula_string, sparse, and tune_grid.

Value

Data frame with simulation results.

See Also

Other simulation functions: plot_simulation

Examples

TODO

[Package hmsPM version 1.0.0 Index]