calc_xgb_weights {hmsPM}	R Documentation

Calculate XGBoost Weights

Description

Calculate XGBoost Weights

Usage

calc_xgb_weights(fit, formula_string, sparse = TRUE, df_test, yhat,
  df_explainer, b_sample = NULL, b_all = NULL, select_class = NULL,
  id_name = "id", top_n = NULL, rest_name = "... the rest",
  digits = 2, verbose = TRUE)

Arguments

fit
Xgb caret fit.

formula_string
Formula string used in caret fit (or for creation of matrix used). Should contain no intercept.

sparse
Use sparse matrix for caret fit if TRUE.

df_test
Data frame (with test data) for which to calculate explanations.

yhat
Predictions (vector) of test data which should be explained.

df_explainer
Explainer data as created by calc_xgb_explainer.

b_sample
Named vector with priori probabilities of sampled data.

b_all
Named vector with priori probabilities of all data.

select_class
In case of multiclass target the class for which explanations should be created (per default the "winner" class is taken).

id_name
Name of id variable in test data (required).

top_n
Only the top n weights per observation are plotted. All "others" are summarized under "... the rest".

rest_name
Label under wich "others" are summarized.

digits
Returned weights are rounded by digits.

verbose
(Logical) Show messages?

Value

Data frame with weights per feature which explain the prediction for all observations of df_test. TODO Columns?

See Also

xgboost

Other xgb explainer functions: calc_shapely_weights, calc_xgb_explainer, plot_all_xgb_explanations, plot_xgb_explanations

[Package hmsPM version 1.0.0 Index]