cate_encoding {hmsPM}	R Documentation

Generate Encoding for Categorical Features

Description

The following encoding methods are mostly suited for tree-based prediction models. This function creates a named vector (lookup table) which can be used to encode feature values with the encode-function.

Usage

cate_encoding(x, method, y, y_ref_level, n_top_levels = 20,
  decreasing = TRUE, enc_nas = TRUE, enc_others = TRUE,
  na_value = NULL, other_value = NULL)

Arguments

x
Feature values to be encoded.

method
Encoding method. Options: label, count, target, self, self_topn, count_others.

y
Target values. Required for target-encoding.

y_ref_level
Target reference level. Required for target-encoding.

n_top_levels
Maximum number of levels. Required for self_topn- and count_others-encoding.

decreasing
Sort order of encoded values. Optional for count-, count_others-, and target-encoding.

enc_nas
(Logical) Encode missing values? Default = TRUE. Required for all encoding methods.

enc_others
(Logical) Encode unkown/other values? Default = TRUE. Required for label-, count-, target-, and self-encoding.

na_value
Replacement value for NAs. Optional for self- and self_topn-encoding.

other_value
Replacement value for unknown/other values to be encoded later. Optional for self- and self_topn-encoding.

Details

Method Specific Arguments

Overview of required and optional paramaters as well as default argument values for all encoding methods.

method 	 label 	 count 	 target 	 self 	 self_topn 	 count_others

y 	 n.a. 	 n.a. 	 required 	 n.a. 	 n.a. 	 n.a.

y_ref_level 	 n.a. 	 n.a. 	 required 	 n.a. 	 n.a. 	 n.a.

n_top_levels 	 n.a. 	 n.a. 	 n.a. 	 n.a. 	 20 	 20

decreasing 	 n.a. 	 TRUE 	 TRUE 	 n.a. 	 n.a. 	 TRUE

enc_nas 	 TRUE 	 TRUE 	 TRUE 	 TRUE 	 TRUE 	 TRUE

enc_others 	 TRUE 	 TRUE 	 TRUE 	 TRUE 	 n.a. 	 n.a.

na_value 	 n.a. 	 n.a. 	 n.a. 	 optional 	 optional 	 n.a.

other_value 	 n.a. 	 n.a. 	 n.a. 	 optional 	 optional 	 n.a.

Method Specific OTHER and NA Encodings

The encoding of OTHER and NA values can be used with the argument values enc_others = TRUE and enc_nas = TRUE, respectively. The other_value and na_value can be specified for the self- and self_topn-encodings. The other encoding methods have predefined values for other_value and na_value.

method 	 OTHER argument 	 other_value 	 NA argument 	 na_value

label 	 n.a. 	 max(enc)+1 	 n.a. 	 0

count 	 n.a. 	 max(enc)+1 	 n.a. 	 0

target 	 n.a. 	 max(enc)+1 	 n.a. 	 0

self 	 other_value 	 _OTHER_ 	 na_value 	 _MISSING_

self_topn 	 other_value 	 _OTHER_ 	 na_value 	 _MISSING_

count_others 	 n.a. 	 max(enc)+1 	 n.a. 	 0

Method Descriptions

label

The feature values are numerated in the order of their first appearance starting at 1.

count

The feature values are numerated according to their frequency in (by default) decreasing order starting at 1.

target

The feature values are encoded with the count-encoded mean of the target reference level y_ref_level over each feature value.

self

Self encoding. This method can be used for OTHER and NA encoding.

self_topn

Self encoding of the top N most frequent feature levels (n_top_levels).

count_others

Count encoding for remaining levels of self_topn-encoding. All n_top_levels are encoded to the value n_levels-n_top_levels+1, all remaining levels are numerated according to their frequency in (by default) decreasing order starting at 1.

Value

Named vector with encoding for feature values.

See Also

Other feature encoding functions: encode

Examples

cate_encoding(c("a", "c", "b"), method = "label")
cate_encoding(letters[c(1:3,2:3,3)], method = "count")
cate_encoding(letters[c(1:3,2:3,3)], method = "self_topn", n_top_levels = 2)
cate_encoding(letters[c(1:3,2:3,3)], method = "count_others", n_top_levels = 2)

[Package hmsPM version 1.0.0 Index]