cate_encoding.self {hmsPM}	R Documentation

Self Encoding

Description

Self encoding. This method can be used for OTHER and NA encoding. The result of self encodings are always factors.

Usage

cate_encoding.self(x, enc_nas = TRUE, enc_others = TRUE,
  na_value = NULL, other_value = NULL)

Arguments

x
Feature values to be encoded.

enc_nas
(Logical) Encode missing values? Default = TRUE.

enc_others
(Logical) Encode unkown/other values? Default = TRUE.

na_value
(Optional) Replacement value for NAs. For metr features, the default value is 0. For cate feature, if the na_value is not specified and x does not contain missing values, the na_value is set to the stat_mode(x).

other_value
(Optional) Replacement value for unknown/other values to be encoded later. For metr features, the default value is median(x). For cate features, if other_value is not specified, it is set to the stat_mode(x).

Value

Named vector with encoding for feature values.

Examples

## Not run: TODO

[Package hmsPM version 1.0.0 Index]