check_feature_scale {hmsPM}	R Documentation

Check feature scale

Description

Check that the feature scale is valid and compare it with the detected feature scale return from detect_feature_scale. Return the detected feature scale if the feature scale is NULL, otherwise the feature scale. If the scales don't match a warning is thrown.

Usage

check_feature_scale(feature_scale, d_feature_scale, feature_name = NULL,
  show_warning = TRUE)

Arguments

feature_scale
Feature scale: metr, ordi, or nomi.

d_feature_scale
Detected feature scale.

feature_name
(Optional) Feature name.

Value

Valid feature scale string.

See Also

Other data check functions: check_same_feature_scales, check_target_type, detect_feature_scale, detect_target_type

Examples

check_feature_scale("nomi", "nomi")
check_feature_scale("ordi", "nomi")

[Package hmsPM version 1.0.0 Index]