check_same_feature_scales {hmsPM}	R Documentation

Check for identical feature scales

Description

Check that all feature scales are valid and the same and throw an error if this is not the case.

Usage

check_same_feature_scales(feature_scales)

Arguments

feature_scales
Vector of feature scales.

Value

feature_scales

See Also

Other data check functions: check_feature_scale, check_target_type, detect_feature_scale, detect_target_type

Examples

check_same_feature_scales(c("nomi", "nomi", "nomi"))
check_same_feature_scales(c("nomi", "nomi", "ordi"))

[Package hmsPM version 1.0.0 Index]