check_target_type {hmsPM}	R Documentation

Check target type

Description

Check that the target type is valid and compare it with the detected target type returned by detect_target_type. Return the detected target type if the target type is NULL, otherwise the target type. If the scales don't match a warning is thrown.

Usage

check_target_type(target_type, d_target_type, target_name = NULL,
  show_warning = TRUE)

Arguments

target_type
Target type: regr, class, or multiclass

d_target_type
Detected target type.

target_name
(Optional) Target name.

Value

Valid feature scale string.

See Also

Other data check functions: check_feature_scale, check_same_feature_scales, detect_feature_scale, detect_target_type

Examples

check_target_type("regr", "regr")
check_target_type("regr", "class")

[Package hmsPM version 1.0.0 Index]