plot_all_performances {hmsPM}	R Documentation

Plot all performance plots

Description

Plot list of performance plots.

Usage

plot_all_performances(y, yhat, reduce_factor = NULL,
  quantiles = seq(0.1, 0.9, 0.1), colors = NULL,
  gradcol = (grDevices::colorRampPalette(hmsPM:::COLORS_HEX))(100),
  ylim = NULL, oneclass = FALSE)

Arguments

y
Target values.

yhat
Vector (Regression) or matrix (Classification) of predicted values as returned by caret's predict function.

reduce_factor
Parameter to thin out ROC curve in case building needs too much memory, e.g., with a value of 0.01 only every hundreds point is taken.

quantiles
Numeric vector with probs of quantiles to use for calibration plot.

colors
Colors for target levels. Used in (multi-)class performance plots.

gradcol
Gradient colors. Used in regression performance and residual plots.

ylim
Y-axis limits (used in several plots).

oneclass
Create some plots only for target class in case of classification.

Details

(Multi-)Classification Plots:

ROC

Confusion Matrix

Distribution

Calibration

Gain

Lift

Precision-Recall

Precision

Regression Plots:

Performance

Distribution

Calibration

Residuals

Value

List of performance plots.

Examples

TODO

[Package hmsPM version 1.0.0 Index]