plot_corr {hmsPM}	R Documentation

Correlation plot

Description

Plot correlation matrix heatmap for variables of same scale.

Usage

plot_corr(df_plot, feature_scale = NULL, cutoff = 0,
  method = "spearman", text_color = "white", device = "ggplot",
  verbose = TRUE)

Arguments

df_plot
Data frame with variables of one scale level / type.

feature_scale
Scale level of variables.

cutoff
Lower cutoff level. Value in [0,1]. Remove rows and columns in which no correlation value is above the cutoff value.

method
Correlation method. Options: 'spearman' or 'pearson'. For categorical data the (corrected) contingency coefficient is calculated.

text_color
Text color.

device
Output device. Options: ggplot or d3heatmap.

verbose
(Logical) Show messages?

Value

Device output object ('ggplot2' or 'd2heatmap' object)

Examples

TODO

[Package hmsPM version 1.0.0 Index]