plot_distr {hmsPM}	R Documentation

Plot distribution

Description

Plot distribution of feature on target

Usage

plot_distr(x, y, feature_name = "feature", target_name = "target",
  feature_scale = NULL, target_type = NULL, varimps = NULL,
  colors = NULL, nbins = NULL, offset = NULL, min_width = NULL,
  inner_barplot = TRUE, levels2extract = NULL, ylim = NULL,
  alpha_fill = 0.2, show_legend = TRUE, varimp_digits = 2,
  verbose = TRUE)

Arguments

x
Feature values.

y
Target values.

feature_name
Feature name.

target_name
Target name.

feature_scale
Feature scale: 'metr', 'ordi', 'nomi'. Automatically detected by detect_feature_scale if not specified.

target_type
Target type: 'regr', 'class', or 'multiclass'. Automatically detected by detect_target_type if not specified.

varimps
Named vector with variable importances with an entry for the value of the 'feature_name' parameter.

colors
(Optional) Colors.

nbins
(Optional) Number of bins. Only relevant for metric distributions.

offset
(Optional) Percentage of plot which is taken to plot inlay plots. Only relevant for metric distributions on (multi-)class targets.

min_width
(Optional) Minimum width of bars for categorical variables. NOT relevant for metric distributions on (multi-)class targets.

inner_barplot
(Logical) Add a distribution barplot for categorical variables. Only relevant for categorical distributions on (multi-)class targets.

levels2extract
Levels to be extracted if at all. Only relevant for categorical distributions on (multi-)class targets.

ylim
(Optional) Y-axis limits. See also lims. NOT relevant for metric distributions on (multi-)class targets.

alpha_fill
Transparency value [0,1] for fill colors. Only relevant for (multi-)class targets.

show_legend
(Logical) Show legend?

varimp_digits
Number of digits for displayed variable importance

verbose
(Logical) Show messages?

Value

Single ggplot.

See Also

Other distribution plots: plot_distr.cate_class, plot_distr.cate_multiclass, plot_distr.cate_regr, plot_distr.metr_class, plot_distr.metr_multiclass, plot_distr.metr_regr

Examples

plot_distr(x = 1:100, y = sample(100, replace = T))
plot_distr(x = 1:100, y = letters[sample(length(letters), size = 100, replace = T)])
plot_distr(x = letters[1:100], y = sample(100, replace = T))

[Package hmsPM version 1.0.0 Index]