plot_distr.cate_class {hmsPM}	R Documentation

Plot categorical distribution on class target

Description

Plot categorical distribution on class target

Usage

plot_distr.cate_class(df_data, feature_name, target_name, varimps = NULL,
  colors, min_width, inner_barplot = TRUE, levels2extract = NULL,
  ylim = NULL, alpha_fill = 0.2, show_legend = TRUE, varimp_digits,
  verbose = TRUE)

Arguments

df_data
Data frame with columns named according to the values of the 'feature_name' and 'target_name' parameters.

feature_name
Feature name.

target_name
Target name.

varimps
Named vector with variable importances with an entry for the value of the 'feature_name' parameter.

colors
Colors for different target values.

min_width
Minimum width of bars.

inner_barplot
(Logical) Add a distribution barplot for categorical variables.

levels2extract
Levels to be extracted if at all.

ylim
Y-axis limits. See also lims.

alpha_fill
Transparency value [0,1] for fill colors.

show_legend
(Logical) Show legend?

varimp_digits
Number of digits for displayed variable importance

verbose
(Logical) Show messages?

Value

Single ggplot.

See Also

Other distribution plots: plot_distr.cate_multiclass, plot_distr.cate_regr, plot_distr.metr_class, plot_distr.metr_multiclass, plot_distr.metr_regr, plot_distr

Examples

TODO

[Package hmsPM version 1.0.0 Index]