plot_distr.metr_class {hmsPM}	R Documentation

Plot metric distribution on multiclass target

Description

Plot metric distribution on multiclass target

Usage

plot_distr.metr_class(df_data, feature_name, target_name, varimps = NULL,
  colors, nbins, offset, alpha_fill = 0.2, show_legend = TRUE,
  varimp_digits = 2, verbose = TRUE)

Arguments

df_data
Data frame with columns named according to the values of the 'feature_name' and 'target_name' parameters.

feature_name
Feature name.

target_name
Target name.

varimps
Named vector with variable importances with an entry for the value of the 'feature_name' parameter.

colors
Colors for different target values.

nbins
Number of bins.

offset
Percentage of plot which is taken to plot inlay plots.

alpha_fill
Transparency value [0,1] for fill colors.

show_legend
(Logical) Show legend?

varimp_digits
Number of digits for displayed variable importance

verbose
(Logical) Show messages?

Value

Single ggplot.

See Also

Other distribution plots: plot_distr.cate_class, plot_distr.cate_multiclass, plot_distr.cate_regr, plot_distr.metr_multiclass, plot_distr.metr_regr, plot_distr

Examples

TODO

[Package hmsPM version 1.0.0 Index]