plot_varimp {hmsPM}	R Documentation

Plot feature importances

Description

Bar plot of (top) feature importances in decreasing order.

Usage

plot_varimp(df_varimp, n_top_features = NULL, colors = c("blue",
  "orange", "red"), df_varimp_cv = NULL, show_cv_lines = TRUE,
  show_cv_confint = TRUE)

Arguments

df_varimp
Data frame with structure as created by calc_varimp_by_permutation.

n_top_features
Number of top features to be plotted.

colors
Colors to use for grouping by color variable.

df_varimp_cv
Data frame with structure as created by calc_varimp_by_permutation but with additional column run distinguishing various cross-validation runs.

show_cv_lines
Show result of each cross validation run as a line if TRUE.

show_cv_confint
Show confidence interval of all cross validation runs if TRUE.

Value

ggplot2 object.

See Also

Other variable importance functions: calc_varimp_by_permutation, plot_varimp_cum

Examples

TODO

[Package hmsPM version 1.0.0 Index]